+++ example/lib/src/models/post_model.dart
@@ -2,35 +2,44 @@ import 'package:flint_dart/model.dart';
 import 'package:flint_dart/schema.dart';
 
 class PostModel extends Model<PostModel> {
-  @override
-  int? id;
   String? title;
   String? subTitle;
+  DateTime? createdAt;
+  DateTime? updatedAt;
+  @override
+  PostModel fromMap(Map<dynamic, dynamic> map) => PostModel()
+    ..title = map['title']?.toString()
+    ..subTitle = map['subTitle']?.toString()
+    ..createdAt = map["created_at"] is String
+        ? DateTime.parse(map["created_at"])
+        : map["created_at"]
+    ..updatedAt = map["updated_at"] is String
+        ? DateTime.parse(map["updated_at"])
+        : map["updated_at"];
+
+  @override
+  Map<String, dynamic> toMap() {
+    return {
+      'title': title,
+      'subTitle': subTitle,
+      "created_at": createdAt?.toIso8601String(),
+      'updated_at': updatedAt?.toIso8601String()
+    };
+  }
 
   @override
   Table get table => Table(
         name: 'post_models',
         columns: [
           Column(
-              name: 'id',
-              type: ColumnType.integer,
-              isPrimaryKey: true,
-              isAutoIncrement: true),
-          Column(name: 'title', type: ColumnType.string),
-          Column(name: 'subTitle', type: ColumnType.string),
+            name: 'title',
+            type: ColumnType.string,
+          ),
+          Column(
+            name: 'subTitle',
+            type: ColumnType.string,
+            isNullable: true,
+          ),
         ],
       );
-
-  @override
-  Map<String, dynamic> toMap() => {
-        'id': id,
-        'title': title,
-      };
-
-  @override
-  PostModel fromMap(Map<dynamic, dynamic> map) {
-    return PostModel()
-      ..id = map['id']
-      ..title = map['title'];
-  }
 }